xfwm4: bigger buttons, top left/right px clickable
Allow the top left and right pixels of maximized windows to be clickable
if there's a window button there, for greater Fitts's Law compliance. The
exception is the window menu button, which needs a minor hack to be sized
correctly.

Somewhat darker inactive button icons, to match window title text.

Add an SVG-based build system to allow colours to be changed easily.