gtk-3.0: Add support for undershoot and overshoot
This commit adds CSS rules to theme undershoot and overshoot. This is
what some applications use to indicate that there is more content below
and/or above the current content viewport.

This is notably used for Nautilus. Without those styles, those classes
are shown as an opaque gray area which looks bad and hides information.